move exec_alu from uops to ops
will use this for const folding in lazy too
